Moved navigation bar configuration to generic subclass
