+++ TMDB/SceneDelegate.swift
@@ -14,7 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let _ = (scene as? UIWindowScene) else { return }
         
-        let svc = TMDBSplitViewController()
+        let svc = UISplitViewController()
         svc.viewControllers = [UINavigationController(rootViewController: Categories()), UINavigationController(rootViewController: MovieList())]
         svc.preferredDisplayMode = .allVisible
         window?.rootViewController = svc
