+++ TMDB/ViewControllers/Categories.swift
@@ -4,7 +4,7 @@
 
 import UIKit
 
-class Categories: UIViewController {
+class Categories: TMDBViewController {
     @IBOutlet private weak var moviesCategoryButton: CategoryButton?
     
     var currentCategory: CategoryButton? {
@@ -16,19 +16,10 @@ class Categories: UIViewController {
         }
     }
     
-    override var preferredStatusBarStyle: UIStatusBarStyle {
-        return .lightContent
-    }
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         title = "Categories"
-        navigationController?.navigationBar.barTintColor = UIColor(red: 39.0 / 255.0, green: 51.0 / 255.0, blue: 63.0 / 255.0, alpha: 1.0)
-        navigationController?.navigationBar.isTranslucent = false
-        navigationController?.navigationBar.titleTextAttributes = [NSAttributedString.Key.foregroundColor:UIColor.white]
-        navigationController?.navigationBar.barStyle = .black
-        
         currentCategory = moviesCategoryButton
     }
