+++ TMDB/ViewControllers/MovieList.swift
@@ -14,7 +14,7 @@ enum FilterType: String, CaseIterable {
     case likes = "vote_count.desc"
 }
 
-class MovieList: UIViewController {
+class MovieList: TMDBViewController {
     @IBOutlet weak var collectionView: UICollectionView?
     @IBOutlet weak var filterSegment: UISegmentedControl?
     @IBOutlet weak var loadingView: UIView?
@@ -29,18 +29,10 @@ class MovieList: UIViewController {
     
     private var currentPage = 1
     
-    override var preferredStatusBarStyle: UIStatusBarStyle {
-        return .lightContent
-    }
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         title = "Movies"
-        navigationController?.navigationBar.barTintColor = UIColor(red: 39.0 / 255.0, green: 51.0 / 255.0, blue: 63.0 / 255.0, alpha: 1.0)
-        navigationController?.navigationBar.isTranslucent = false
-        navigationController?.navigationBar.titleTextAttributes = [NSAttributedString.Key.foregroundColor:UIColor.white]
-        navigationController?.navigationBar.barStyle = .black
         
         loadingIndicator?.lineWidth = 4.0
         loadingIndicator?.color = .white
