--- TMDB/ViewControllers/TMDBSplitViewController.swift
@@ -1,15 +1,0 @@
-// Created for TMDB in 2020
-// Using Swift 5.0
-// Running on macOS 10.15
-
-import UIKit
-
-class TMDBSplitViewController: UISplitViewController {
-    override var preferredStatusBarStyle: UIStatusBarStyle {
-        return .lightContent
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
