+++ TMDB/ViewControllers/TMDBViewController.swift
@@ -1,0 +1,20 @@
+// Created for TMDB in 2020
+// Using Swift 5.0
+// Running on macOS 10.15
+
+import UIKit
+
+class TMDBViewController: UIViewController {
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return .lightContent
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        navigationController?.navigationBar.barTintColor = UIColor(red: 39.0 / 255.0, green: 51.0 / 255.0, blue: 63.0 / 255.0, alpha: 1.0)
+        navigationController?.navigationBar.isTranslucent = false
+        navigationController?.navigationBar.titleTextAttributes = [NSAttributedString.Key.foregroundColor:UIColor.white]
+        navigationController?.navigationBar.barStyle = .black
+    }
+}
